Create utility for arrow graph projection in tests
