+++ graphdatascience/tests/integrationV2/procedure_surface/arrow/graph_creation_helper.py
@@ -1,0 +1,33 @@
+import json
+from typing import Optional, Tuple
+
+from graphdatascience import Graph
+from graphdatascience.arrow_client.authenticated_flight_client import AuthenticatedArrowClient
+from graphdatascience.arrow_client.v2.job_client import JobClient
+
+
+class MockGraph(Graph):
+    def __init__(self, name: str):
+        self._name = name
+
+    def name(self) -> str:
+        return self._name
+
+
+def create_graph(
+    arrow_client: AuthenticatedArrowClient, graph_name: str, gdl: str, undirected: Optional[Tuple[str, str]] = None
+) -> Graph:
+    arrow_client.do_action("v2/graph.fromGDL", json.dumps({"graphName": graph_name, "gdlGraph": gdl}).encode("utf-8"))
+
+    if undirected is not None:
+        JobClient.run_job_and_wait(
+            arrow_client,
+            "v2/graph.relationships.toUndirected",
+            {"graphName": graph_name, "relationshipType": undirected[0], "mutateRelationshipType": undirected[1]},
+        )
+        arrow_client.do_action(
+            "v2/graph.relationships.drop",
+            json.dumps({"graphName": graph_name, "relationshipType": undirected[0]}).encode("utf-8"),
+        )
+
+    return MockGraph(graph_name)
